+++ src/main/kotlin/kr/hhplus/be/server/infrastructure/config/RedisConfig.kt
@@ -1,14 +1,9 @@
 package kr.hhplus.be.server.infrastructure.config
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.SerializationFeature
-import com.fasterxml.jackson.databind.jsontype.BasicPolymorphicTypeValidator
-import com.fasterxml.jackson.module.kotlin.KotlinModule
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.redis.connection.RedisConnectionFactory
 import org.springframework.data.redis.core.RedisTemplate
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer
 import org.springframework.data.redis.serializer.StringRedisSerializer
 
 @Configuration
@@ -16,23 +11,14 @@ class RedisConfig {
     @Bean
     fun redisTemplate(
         connectionFactory: RedisConnectionFactory,
-        objectMapper: ObjectMapper
-    ): RedisTemplate<String, Any> {
-        val template = RedisTemplate<String, Any>()
+    ): RedisTemplate<String, String> {
+        val template = RedisTemplate<String, String>()
+        val stringRedisSerializer = StringRedisSerializer()
         template.connectionFactory = connectionFactory
-        template.keySerializer = StringRedisSerializer()
-        template.valueSerializer = Jackson2JsonRedisSerializer(objectMapper, Any::class.java)
+        template.keySerializer = stringRedisSerializer
+        template.valueSerializer = stringRedisSerializer
+        template.hashKeySerializer = stringRedisSerializer
+        template.hashValueSerializer = stringRedisSerializer
         return template
     }
-
-    @Bean
-    fun objectMapper(): ObjectMapper = ObjectMapper()
-        .registerModule(KotlinModule.Builder().build())
-        .activateDefaultTyping(
-            BasicPolymorphicTypeValidator.builder()
-                .allowIfSubType(Any::class.java)
-                .build(),
-            ObjectMapper.DefaultTyping.EVERYTHING
-        )
-        .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
 }
